+++ src/components/generator/dnd/DndContainer.tsx
@@ -3,20 +3,24 @@ import '@atlaskit/css-reset';
 import { v4 as uuid } from 'uuid';
 import TextComponentEditor from '../editor/TextComponentEditor';
 import ImgComponentEditor from '../editor/ImgComponentEditor';
-import { TextComponentType, ImgComponentType } from 'types';
 import styled from '@emotion/styled';
+import { useRecoilState } from 'recoil';
+import { componentsState } from 'atoms/components';
+import DndComponent from './DndComponent';
 
-const DndContainer = ({ post, setPost }: any) => {
+const DndContainer = () => {
+  const [components, setComponents] = useRecoilState(componentsState);
+
   const handleChange = (result: any) => {
     if (!result.destination) return;
-    const items = [...post];
+    const items = [...components];
     const [reorderedItem] = items.splice(result.source.index, 1);
     items.splice(result.destination.index, 0, reorderedItem);
-    setPost(items);
+    setComponents(items);
   };
 
   const addTextEditor = () => {
-    setPost((oldComponents: any[]) => [
+    setComponents((oldComponents: any[]) => [
       ...oldComponents,
       {
         id: `${uuid()}`,
@@ -26,40 +30,13 @@ const DndContainer = ({ post, setPost }: any) => {
     ]);
   };
 
-  const switchComponent = (component: any) => {
-    switch (component.type) {
-      case 'text':
-        return (
-          <TextComponentEditor key={component.id} textComponent={component as TextComponentType} />
-        );
-      case 'img':
-        return (
-          <ImgComponentEditor key={component.id} imgComponent={component as ImgComponentType} />
-        );
-      default:
-        return;
-    }
-  };
-
   return (
     <DragDropContext onDragEnd={handleChange}>
       <Droppable droppableId="cardlists">
         {provided => (
           <div className="cardlists" {...provided.droppableProps} ref={provided.innerRef}>
-            {post.map((e: any, i: number) => (
-              <Draggable draggableId={`test-${e.id}`} index={i} key={`test-${e.id}`}>
-                {(provided, snapshot) => {
-                  return (
-                    <div
-                      {...provided.draggableProps}
-                      {...provided.dragHandleProps}
-                      ref={provided.innerRef}
-                    >
-                      {switchComponent(e)}
-                    </div>
-                  );
-                }}
-              </Draggable>
+            {components.map((post: any, i: number) => (
+              <DndComponent key={`test-${post.id}`} post={post} postIndex={i} />
             ))}
             {provided.placeholder}
           </div>
